回転の中心を修正してちゃんと回るように修正

+++ ScreenForm.cs
@@ -72,11 +72,11 @@ namespace KamopisScreensaver
                 var vx = KamopisBaseSize / 10 * (rand.NextDouble() - 0.5) + 5;
                 var vy = KamopisBaseSize / 10 * (rand.NextDouble() - 0.5) + 5;
 
-                var t = 360 * rand.NextDouble();            // 初期角度
-                var vt = 20 * (rand.NextDouble() - 0.5);    // 角加速度
+                var t = 360 * rand.NextDouble();
+                var vt = 20 * (rand.NextDouble() - 0.5);
 
                 var s = KamopisBaseSize * (1 + (rand.NextDouble() * KAMOPIS_SIZE_RANGE - KAMOPIS_SIZE_RANGE / 2));
-                this.Kamopises.Add(new Kamopis(x, y, vx, vy, t, vt, s, s));
+                this.Kamopises.Add(new Kamopis((float)x, (float)y, (float)vx, (float)vy, (float)t, (float)vt, (float)s, (float)s));
             }
         }
